requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.3
boto3==1.40.55
botocore==1.40.55
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dnspython==2.8.0
Faker==37.4.0
fastapi==0.115.12
fonttools==4.57.0
frozenlist==1.5.0
google==3.0.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-oauth2-tool==0.0.3
googleapis-common-protos==1.69.2
greenlet==3.2.1
gspread==6.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.1
MouseInfo==0.1.3
mss==10.0.0
multidict==6.3.2
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandasql==0.7.3
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pydantic==2.11.3
pydantic_core==2.33.1
PyGetWindow==0.0.9
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.14.0
selenium==4.31.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
whatsapp-api-client-python==0.0.49
whatsapp-python==4.3.0
wheel==0.45.1
wsproto==1.2.0
yarl==1.19.0
